requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
avalanche-lib==0.1.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Cython==0.29.30
docker-pycreds==0.4.0
filelock==3.4.2
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gdown==4.2.1
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.6.0
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.43.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.10.1
joblib==1.1.0
kiwisolver==1.3.2
lime
markdown==3.3.6
matplotlib==3.5.1
multidict==6.0.2
networkx==2.6.3
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
path==16.4.0
pathtools==0.1.2
Pillow==9.0.1
problog==2.2.2
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyDeprecate==0.3.1
pyparsing==3.0.7
pysd==3.2.0
PySDD==0.2.10
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.9
pytorchcv==0.0.67
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
quadprog==0.1.11
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.5.4
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.10.2
torchmetrics==0.7.0
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.10
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
